requirements.txt
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
billiard==4.2.4
cachetools==6.2.2
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
exceptiongroup==1.3.1
fastapi==0.122.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
greenlet==3.2.4
h11==0.16.0
httplib2==0.31.0
idna==3.11
kombu==5.6.1
mako==1.3.10
markupsafe==3.0.3
packaging==25.0
prompt-toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.5
pydantic-core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.21
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.50.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.1
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
